[hail] move most ReferenceGenome, liftover functions to single LocusFunctions object
