+++ hail/python/hail/genetics/reference_genome.py
@@ -494,7 +494,6 @@ class ReferenceGenome(object):
         if dest_reference_genome.name in self._liftovers:
             raise KeyError(f"Liftover already exists from {self.name} to {dest_reference_genome.name}.")
         self._liftovers[dest_reference_genome.name] = chain_file
-        hl.ir.register_liftover_functions(self.name, dest_reference_genome.name)
 
 
 rg_type.set(ReferenceGenome)
