+++ hail/python/hail/ir/register_functions.py
@@ -6,28 +6,11 @@ from .ir import register_function, register_session_function, register_seeded_fu
 def register_reference_genome_functions(rg):
     from hail.expr.types import dtype
 
-    tvariant = dtype(f"struct{{locus:locus<{rg}>,alleles:array<str>}}")
-    tinterval = dtype(f"interval<locus<{rg}>>")
-
-    register_session_function(f"Locus({rg})", (dtype("str"),), dtype(f"locus<{rg}>"))
-    register_session_function(f"Locus({rg})", (dtype("str"),dtype("int32"),), dtype(f"locus<{rg}>"))
-    register_session_function(f"LocusAlleles({rg})", (dtype("str"),), tvariant)
-    register_session_function(f"LocusInterval({rg})", (dtype("str"),dtype("bool"),), tinterval)
-    register_session_function(f"LocusInterval({rg})", (dtype("str"),dtype("int32"),dtype("int32"),dtype("bool"),dtype("bool"),dtype("bool"),), tinterval)
     register_session_function(f"isValidContig({rg})", (dtype("str"),), dtype("bool"))
     register_session_function(f"isValidLocus({rg})", (dtype("str"),dtype("int32"),), dtype("bool"))
 
     register_session_function(f"getReferenceSequenceFromValidLocus({rg})", (dtype("str"),dtype("int32"),dtype("int32"),dtype("int32"),), dtype("str"))
     register_session_function(f"getReferenceSequence({rg})", (dtype("str"),dtype("int32"),dtype("int32"),dtype("int32"),), dtype("str"))
-
-    register_session_function(f"globalPosToLocus({rg})", (dtype("int64"),), dtype(f"locus<{rg}>"))
-    register_session_function(f"locusToGlobalPos({rg})", (dtype(f"locus<{rg}>"),), dtype("int64"))
-
-def register_liftover_functions(rg, dest_rg):
-    from hail.expr.types import dtype
-
-    register_session_function(f"liftoverLocus({rg})({dest_rg})", (dtype(f"locus<{rg}>"), dtype('float64'),), dtype(f"struct{{result:locus<{dest_rg}>,is_negative_strand:bool}}"))
-    register_session_function(f"liftoverLocusInterval({rg})({dest_rg})", (dtype(f"interval<locus<{rg}>>"), dtype('float64'),), dtype(f"struct{{result:interval<locus<{dest_rg}>>,is_negative_strand:bool}}"))
 
 
 def register_functions():
@@ -211,6 +194,15 @@ def register_functions():
     register_function("toFloat64", (dtype("str"),), dtype("float64"))
     register_function("toFloat64", (dtype("bool"),), dtype("float64"))
     register_function("dbeta", (dtype("float64"),dtype("float64"),dtype("float64"),), dtype("float64"))
+    register_function(f"Locus", (dtype("str"),), dtype("?T:locus"))
+    register_function(f"Locus", (dtype("str"), dtype("int32"),), dtype("?T:locus"))
+    register_function(f"LocusAlleles", (dtype("str"),), dtype("struct{locus: ?T, alleles: array<str>}"))
+    register_function(f"LocusInterval", (dtype("str"),dtype("bool"),), dtype("interval<?T:locus>"))
+    register_function(f"LocusInterval", (dtype("str"),dtype("int32"),dtype("int32"),dtype("bool"),dtype("bool"),dtype("bool"),), dtype("interval<?T:locus>"))
+    register_function(f"globalPosToLocus", (dtype("int64"),), dtype("?T:locus"))
+    register_function(f"locusToGlobalPos", (dtype("?T:locus"),), dtype("int64"))
+    register_function(f"liftoverLocus", (dtype(f"?T:locus"), dtype('float64'),), dtype(f"struct{{result:?U:locus,is_negative_strand:bool}}"))
+    register_function(f"liftoverLocusInterval", (dtype(f"interval<?T:locus>"), dtype('float64'),), dtype(f"struct{{result:interval<?U:locus>,is_negative_strand:bool}}"))
     register_function("min_rep", (dtype("?T:locus"),dtype("array<str>"),), dtype("struct{locus: ?T, alleles: array<str>}"))
     register_function("locus_windows_per_contig", (dtype("array<array<float64>>"),dtype("float64"),), dtype("tuple(array<int32>, array<int32>)"))
     register_function("toBoolean", (dtype("str"),), dtype("bool"))
